+++ lib/src/framework/provide_it_caching.dart
@@ -10,29 +10,21 @@ extension on ReadItMixin {
     return 0;
   }
 
-  _State? _stateOf<T>(BuildContext? context, {String? type, Object? key}) {
-    type ??= T.type;
-
-    final state = getRefStateOfType<T>(type: type, key: key);
+  _State? _stateOf<T>(BuildContext? context, {Object? key}) {
+    final state = getRefStateOfType<T>(key: key);
 
     if (key case Ref<T> ref) {
       final bind = (element: context, index: _treeIndex[context] ?? 0);
+      // get-or-bind
       if (state == null || state._bind == bind) {
-        return _state<T>(context as Element?, ref, true);
+        return _state<T>(context as Element?, ref);
       }
     }
 
-    if (state?.type == type && context is Element) {
+    if (state != null && context is Element) {
       final index = _dependencyIndex[context] ??= _initCacheIndex(context);
       _dependencyIndex[context] = index + 1;
-
-      return state;
     }
-
-    // case it's a global ref, we can auto-bind it.
-    // if (key case Ref<T> ref) {
-    //   return _state(context as Element?, ref);
-    // }
 
     return state;
   }