+++ lib/src/framework/read_it.dart
@@ -12,7 +12,7 @@ sealed class ReadIt {
   static final I = instance;
 
   /// Binds a [ProvideRef] without context.
-  R bind<R, T>(ProvideRef<T> ref);
+  RefState<T, Ref<T>> bind<T>(ProvideRef<T> ref);
 
   /// Reads the value of a [Ref].
   T read<T>({Object? key});
@@ -56,7 +56,7 @@ extension ReadItProviders on ReadIt {
     return bind(ProvideRef.value(
       value,
       key: key,
-    ));
+    )).read();
   }
 }